api/requirements.txt
pandas
requests
google-cloud-bigquery

